20160630_2028 (Hangman - fixed db issue)
Where clause in retrieving the database is what was causing the cursor
to fail.  Need to figure out what the problem with the where clause is.

Also moved creating the prefs manager to the constructor so it only
pulls the info from the preference once rather than every time
nextName() is called.