20160722_1425 (Hangman - handle when names left are less than 10)
Checks if nextTenNamesFromDB() is returning less than 10 names - if it
does, it pads empty names to the front of the ArrayList and increases
nameReturnedCounter so nextTenNamesFromDB() will be called again as
soon as those less than 10 names are displayed.

Added a blank constructor to HangmanNameItem class

Also added a variable that keeps track of the number of names in the
database (namesInDB) and added a comment to the loaddatabasetask class
that instructs whoever is adding new names to update the variable with
a new literal.